PNDA-196 Moved Storage classes to no longer exist in gcloud.
