+++ aws/pom.xml
@@ -67,7 +67,7 @@
     <dependency>
     	<groupId>org.jimmutable</groupId>
     	<artifactId>core</artifactId>
-    	<version>0.0.7</version>
+    	<version>0.0.7-SNAPSHOT</version>
     </dependency>
     
     <dependency>
