+++ aws/src/main/java/org/jimmutable/storage/StorageKeyExtension.java
@@ -1,4 +1,4 @@
-package org.jimmutable.gcloud;
+package org.jimmutable.storage;
 
 import org.jimmutable.core.exceptions.ValidationException;
 import org.jimmutable.core.objects.Stringable;
